Docker deployment enabled in API

+++ src/main/java/com/shortly/Repository/UrlRepo.java
@@ -12,7 +12,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface UrlRepo extends JpaRepository<UrlMap, Integer> {
+public interface UrlRepo extends JpaRepository<UrlMap, Long> {
 
     @Query("select url from UrlMap url where url.isActive = :is_active and url.shortCode = :shortCode and (url.expiresAt is null OR url.expiresAt > :currentDate)")
     Optional<UrlMap> findUrlData(@Param("shortCode") String shortUrl,@Param("is_active") boolean is_active,@Param("currentDate") Date currentDate);
@@ -26,7 +26,7 @@ public interface UrlRepo extends JpaRepository<UrlMap, Integer> {
 
     List<UrlMap> findByUserUsername(String userName);
 
-    Optional<UrlMap> findByIdAndUserUsername(Integer id, String userName);
+    Optional<UrlMap> findByIdAndUserUsername(Long id, String userName);
 
     Optional<UrlMap> findByShortCodeAndIsActiveAndUserUsername(String shortCode, boolean b, String username);
